OCV data update
The OCV table data has been updated with values obtained from a MATLAB simulation, which are more accurate to reality. The method of comparing the sensor voltage with the table data has been reverted to a for loop because the voltage step in the simulation data is not uniform. The old comparison method has been left in the comments in the code in case it needs to be switched back in the future.